backend/requirements.txt
fastapi
python-dateutil
python-docx
python-multipart
pymupdf
requests
sqlalchemy
uvicorn